requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.9
requests==2.32.3
rerequests==1.0.0b0
sqlparse==0.5.0
urllib3==2.2.2
